+++ packages/suite-web/e2e/tests/settings/coins.test.ts
@@ -27,7 +27,7 @@ describe('Coin Settings', () => {
             'zec',
             'ada',
             'test',
-            'trop',
+            'tgor',
             'txrp',
             'tada',
         ];