+++ packages/theme/src/coinsColors.ts
@@ -17,7 +17,6 @@ export const coinsColors = {
     zec: '#F5B300',
     btc: '#F29937',
     test: '#E75F5F',
-    trop: '#E75F5F',
     txrp: '#E75F5F',
     gnt: '#8A92B2',
     ada: '#3468D1',