+++ suite-common/wallet-utils/src/__tests__/accountUtils.test.ts
@@ -153,7 +153,6 @@ describe('account utils', () => {
 
     it('isTestnet', () => {
         expect(isTestnet('test')).toEqual(true);
-        expect(isTestnet('trop')).toEqual(true);
         expect(isTestnet('tgor')).toEqual(true);
         expect(isTestnet('txrp')).toEqual(true);
         expect(isTestnet('btc')).toEqual(false);