+++ suite-common/wallet-utils/src/accountUtils.ts
@@ -134,8 +134,6 @@ export const getTitleForNetwork = (symbol: NetworkSymbol) => {
             return 'TR_NETWORK_ZCASH';
         case 'eth':
             return 'TR_NETWORK_ETHEREUM';
-        case 'trop':
-            return 'TR_NETWORK_ETHEREUM_TESTNET';
         case 'tgor':
             return 'TR_NETWORK_ETHEREUM_GOERLI';
         case 'etc':