+++ suite-common/wallet-utils/src/validation.ts
@@ -4,7 +4,7 @@ import { Account } from '@suite-common/wallet-types';
 
 const getNetworkType = (symbol: Account['symbol']) => {
     if (symbol === 'regtest') return symbol;
-    const testnets = ['test', 'txrp', 'trop', 'tgor', 'tada'];
+    const testnets = ['test', 'txrp', 'tgor', 'tada'];
     return testnets.includes(symbol) ? 'testnet' : 'prod';
 };
 
@@ -17,7 +17,6 @@ const getCoinFromTestnet = (symbol: Account['symbol']) => {
             return 'xrp';
         case 'tada':
             return 'ada';
-        case 'trop':
         case 'tgor':
             return 'eth';
         default: